requirements.txt
# flask==2.0.1
flask
numpy==1.21.1
# opencv==4.5.0
opencv
# pandas==2.0.0
pandas
# pillow==8.3.1
pillow
# python==3.8.0
# waitress==2.1.2
# waitress
gunicorn


